+++ src/main/scala/onion/compiler/SemanticError.scala
@@ -135,6 +135,7 @@ object SemanticError {
   case object MISSING_RETURN extends SemanticError(67)
   case object OVERRIDE_TARGET_NOT_FOUND extends SemanticError(68)
   case object VAL_REQUIRES_INITIALIZER extends SemanticError(69)
+  case object NULLABLE_MEMBER_ACCESS extends SemanticError(70)
 }
 sealed abstract class SemanticError(val code: Int) {
   /** Returns the error code in format "E0001" */